Refactor the canvas test cases into their own widget
